Refactor post-carreras component and add navigation after successful creation
Refactor the post-carreras component to include navigation to the admin/inicio route after successfully creating a new carrera. This improves the user experience by automatically redirecting them to the appropriate page.

Fix typo in eliminarCarrera function and update modal title

Fix a typo in the eliminarCarrera function and update the modal title in the tablero-admin component. The modal title now correctly displays the name of the carrera being deleted.

Update eliminarCarrera function to use selectedCarrera object

Update the eliminarCarrera function in the tablero-admin component to use the selectedCarrera object instead of passing the id and nombre as separate parameters. This simplifies the code and improves readability.

Update sidebar component to remove unnecessary query parameters

Update the sidebar component to remove the unnecessary query parameters in the routerLink for the Licenciaturas button. This improves the code by removing unnecessary complexity.